refactor: コマンド募集とモーダル募集の型、クラスを共通化
