+++ src/app/feat-recruit/types/recruit_condition_error.ts
@@ -1,0 +1,13 @@
+import { ErrorTexts } from '../../constant/error_texts';
+
+export class RecruitConditionError extends Error {
+    private errorMessage: string | undefined;
+    constructor(errorMessage?: string) {
+        super();
+        this.errorMessage = errorMessage;
+    }
+
+    public getErrorMessage() {
+        return this.errorMessage ?? ErrorTexts.UndefinedError;
+    }
+}
